Conflicts:
	JHUGenerator/main.F90
	JHUGenerator/mod_Kinematics.F90
	JHUGenerator/mod_Parameters.F90

+++ JHUGenMELA/pdfs/downloadNNPDF.sh
@@ -1,11 +1,24 @@
 #!/bin/bash
 
-# wget http://nnpdf.hepforge.org/html/nnpdf23/PDFsets/NNPDF23_lo_as_0130.LHgrid.tgz
-wget http://pcteserver.mi.infn.it/~nnpdf/nnpdf30/NNPDF30_lo_as_0130.LHgrid.tgz
+if ! [ $(pwd | sed "s|.*/||") == pdfs ]; then
+    echo "Please run downloadNNPDF from the pdfs directory!"
+    exit 1
+fi
 
-# tar -zxvf ./NNPDF23_lo_as_0130.LHgrid.tgz
-tar -zxvf ./NNPDF30_lo_as_0130.LHgrid.tgz
+if [ -f NNPDF30_lo_as_0130.LHgrid ]; then
+    exit 0
+fi
 
-# rm ./NNPDF23_lo_as_0130.LHgrid.tgz
-rm ./NNPDF30_lo_as_0130.LHgrid.tgz
+if [ -f ../../JHUGenerator/pdfs/NNPDF30_lo_as_0130.LHgrid ]; then
+    ln -s ../../JHUGenerator/pdfs/NNPDF30_lo_as_0130.LHgrid .
+else
+    # wget http://nnpdf.hepforge.org/html/nnpdf23/PDFsets/NNPDF23_lo_as_0130.LHgrid.tgz
+    wget http://pcteserver.mi.infn.it/~nnpdf/nnpdf30/NNPDF30_lo_as_0130.LHgrid.tgz
+
+    # tar -zxvf ./NNPDF23_lo_as_0130.LHgrid.tgz
+    tar -zxvf ./NNPDF30_lo_as_0130.LHgrid.tgz
+
+    # rm ./NNPDF23_lo_as_0130.LHgrid.tgz
+    rm ./NNPDF30_lo_as_0130.LHgrid.tgz
+fi
